[hyperactor] introduce typed ref_ capability layer
Extract the typed capability reference surface into a real `ref_` module. Move `ActorRef`, `PortRef`, `OncePortRef`, and the unbound-port helpers there, and keep `reference` as a compatibility adapter that re-exports the typed layer while it still owns the legacy ID adapters.

Differential Revision: [D102664682](https://our.internmc.facebook.com/intern/diff/D102664682/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D102664682/)!

[ghstack-poisoned]
